finished project

+++ app/Http/Controllers/ListingController.php
@@ -33,6 +33,10 @@ class ListingController extends Controller
     //Update Listing Data
     public function update(Request $request, Listing $listing)
     {
+        //Make sure logged in User is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
         // dd($request->all());
         $formFields = $request->validate([
             'title' => 'required',
@@ -78,6 +82,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
 
@@ -87,7 +93,18 @@ class ListingController extends Controller
     //Delete Listing
     public function destroy(Listing $listing)
     {
+        //Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing Deleted Successfully');
+    }
+
+    //Manage Listings
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
